prost: no recursion limit for compute connection
Prost has a default recursion limit of 100, meaning that it fails to decode
any message that has a nesting level of more than 100. In most scenarios
this should be a reasonable default, but the way the dataflow plan is
encoded, we can easily exceed this limit when creating large join plans.

This change disables the recusion limit for prost for the `compute-client`
crate. While this is not a great fix, it avoids the problem for the moment.

In the future, we should think about a different representation of plans,
which use a flat and/or stack-based representation. Both could help limit
the depth of the message structure.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>