storage: allow multiple sources to share storaged
This commit introduces the concept of a "storage host", which is a
storaged process that can host multiple "storage objects" (either a
source or, soon, a sink). While we don't plan to expose this
functionality to users in the short term, it is useful in test
environments, where sources can opt in to sharing a storaged process via
the `REMOTE` option. In the long term, this functionality may prove
useful in production if we adopt some scheduling smarts to pack multiple
sources and sinks onto a single storage host.

The responsibility for creating storage hosts and assigning storage
objects to them lies with a new `StorageHosts` object. This object is
largely bookkeeping over some hash maps to ensure the controller opens
only one connection to a given `storaged` process.

Fix #13367.