persist: remove old Metrics, BlobCache, Maintainer
We kept these around during the old persist deletion in case they came
in handy and to possibly retain the git history. Metrics and Maintainer
(now Compactor) have been rewritten since then (it turned out to be
easier/necessary to rewrite). We've also finished porting all the old
persist compaction tests.

There are no immediate plans to reintroduce a BlobCache (maybe for some
future impl of TraceReader?) and even if we did, that'd be rewritten
too.

There is still value in the bounded memory usage technique of old
compaction, but it's just as easy to pull it out of git history.

As a result, remove all this code, which allows for easier iteration on
new persist.