coord: Implement consistent timestamp recovery (Percolator) (#13385)
The consistency of certain queries relies on the Coordinator to return
strictly increasing timestamps. The Coordinator returns timestamps with
the help of a global timestamp. Previously, when the Coordinator
started it initialized the global timestamp to the current system
clock. This is fine if the system clock always goes forward, but this
is not always the case. The system clock can go backwards or the
Coordinator can be restarted on a separate machine with a different
clock. As a consequence, it was possible for the global timestamp to
decrease after a restart.

This commit implements a timestamp recovery protocol that is based on
Percolator timestamp recovery (https://research.google/pubs/pub36726/).
When serving a timestamp, we "pre-allocate" a group of timestamps at
once, and only durably store the largest of those timestamps. All
timestamps within that interval can be served directly from memory,
without going to disk. On restart, we re-initialize the current
timestamp to a value one larger than the persisted timestamp.

Works toward resolving #11631