controller: push `ready`/`process` methods into each controller
The `Controller::ready` and `Controller::process` methods have proven to
be a useful tool to work around cancellation safety issues. The idea is
that `ready` is cancellation safe, waits for work to become available,
then stashes it away for later; when readiness is observed, the
cancellation-unsafe `Controller:process` method must be driven to
completion.

This commit pushes the pattern into `StorageController` and
`ComputeController`. The top-level `Controller`'s `process` and
`ready` methods now delegate to the underlying controllers.

This change avoids leaking implementation details to the `Controller`.
For example, previously, the top-level `Controller` was responsible for
processing `FrontierUppers` responses by calling
`update_write_frontiers`. After this commit, frontier handling occurs
locally to each controller.

The unused `LinearizedTimestamps` response was frustrating this effort,
so I removed it. It would have needed a refactoring anyway, since our
latest thinking did not involve sending the command all the way down to
the storage hosts themselves.

With this change, I am quite comfortable with our cancellation safety
story. There are only a few `GenericClient::recv` and
`Controller::ready` methods that need to be audited for cancellation
safety issues, and they are now all quite simple.

h/t to @brennanvincent for inventing the `ready`/`process` technique.

Touches #12946.