coord: Maintain a timestamp oracle per timeline (#13510)
Previously, the Coordinator only had a single timestamp oracle that
provided timestamps for user and system tables on the EpochMilliseconds
timeline. All reads to user and system tables would request a timestamp
from the global timestamp oracle. In order to provide strict
serializable reads, the Coordinator will have to consult some timestamp
oracle for all reads, not just reads to tables. Reads on objects
that are not in the EpochMilliseconds timeline need to get a timestamp
from a timestamp oracle specific to their timeline.

This commit updates the Coordinator so that it maintains a map of
timestamp oracles, so each timeline has their own timestamp oracle.
Non-realtime timestamp oracles are periodically pushed forward based
on the uppers of all objects in that timeline.

This commit may not reflect how we want timelines to look in the
future, but it does reflect how timelines look right now. The code is
also mostly plumbing. Non-realtime timestamp oracles are maintained and
kept up to date, but they're never actually used.

Works towards resolving #12309