repr: Add `C` parameter to `Display$Format` types
The `fmt_$format` might need to share and modify some
context that asserts that a sub-part (such as a field)
is rendered correctly within the enclosing parent.

To achieve that, all `Display$Format` traits now accept
a generic type called `C` which defaults to `()` (the
empty context).

Individual `Display$Format` implementations will then
fix the expected context type on a case-by-case basis,
defaulting to the empty context for top-level structures.