persist: finish wiring up seqno-capability based blob gc
This registers SnapshotIter and Listen so they get capabilities on since
and seqno (by internally storing a ReadHandle on each). As they
progress, they heartbeat to maintain the registration (so their leases
aren't expired from under them, once we add leases) and keep their
capabilities.

When SnapshotIter finishes, it expires its ReadHandle, giving up all
capabilities. Listen downgrade its since and seqno capability as it
goes.

We make sure to do the SnapshotIter registration before it's sent over
the wire so there isn't a race. This means we have to add a reader_id
field to ProtoSnapshotSplit. This is a backward incompatible change, but
it's fine because it's never durably persisted and we're not doing
rolling upgrades yet. This also adds last_heartbeat_timestamp_ms field
to ProtoReaderState and renames a couple proto messages, all of which
are backward compatible changes.