persist: panic instead of silent incorrectness on invalid usage
Before this commit, if a shard's since was advanced beyond the frontier
of some active listener, that listener could emit incorrect data
(compaction would also have to be quite prompt). This would be an
invalid usage by the persist user (the since shouldn't have been
advanced that far), but it's better to panic then silently emit
incorrect data. This doesn't seem to be happening in practice.

Not bothering with a regression test because the way I found this was an
upcoming branch that causes persist unit tests to deterministically
fail.