sql,coord: remove CREATE MATERIALIZED VIEW[S]
This commit removes support for typing `CREATE MATERIALIZED VIEW` to
create a view and a default index on it at the same time from both the
SQL layer and the coordinator.

The syntax will be added back again in a later commit, when it will be
used for creating what is now called a recorded view.