dbt-materialize: adapt to new materialized views
This commit adjusts the dbt-materialize adapter to work with the new
materialized views, which are now separate database objects, instead of
regular views with an index. Consequently, any macros that previously
assumed we were always dealing with views (or potentially tables) only
need to be adjusted to also handle materialized views correctly.