Persisting Introspection Dataflows Part 1 (#13340)
This commit adds support for writing each replicas introspection source to persists.
Each replica uses a separate shard for each introspection sources. They are then
exposed in the catalog using a replica id as postfix. For example mz_active_peeks
of replica id 3 can be accessed using mz_catalog.mz_active_peeks_3.

The persisted shards are stored in a catalog and re-used on a coordinator restart. The computed's
flush their shard before writing the updated information.
It refactors ConcreteComputeInstanceReplicaConfig to contain also the persisted logs. The former ConcreteComputeInstanceReplicaConfig is now a member and has been renamed to ConcreteComputeInstanceReplicaLocation. A corresponding migration has been implemented.

In an empty catalog, the persisted_logs of the default cluster are set to Default. This value is then
on first boot replaced with an allocated default set of shards and written back to the catalog storage.

Co-authored-by: Moritz Hoffmann <antiguru@gmail.com>

Co-authored-by: Jan Teske <jteske@posteo.net>