storage: don't persist entire CollectionMetadata
There are two fields of `CollectionMetadata` that ought not be persisted
to the storage controller stash:

  * persist_location, which is supplied at runtime and could
    legitimately change during e.g. an operator migration of
    the persist consensus implementation.

  * status_shard, which is derived from the ID of the status collection
    that is provided by the adapter layer.

This commit creates a new type, called `DurableCollectionMetadata`, for
the subset of `CollectionMetadata` that ought to be persisted to disk.
The controller is responsible for converting a
`DurableCollectionMetadata` into a `CollectionMetadata` on boot by
deriving the necessary fields.

The `ProtoDurableCollectionMetadata` type has the same on-disk encoding
as `ProtoCollectionMetadata`, except with the unnecessary fields
omitted, so no migration is necessary. This leaves the unnecessary
fields on disk for existing stash entries; that seems okay because we'll
be doing at least one or two full wipes in the next few weeks.

Supersedes #13805 by removing the field that would have required the
migration.