coord: Split coord.rs into multiple modules (#13362)
This commit attempts to split coord.rs into multiple sub-modules.
coord.rs has become extremely large and close to unmanageable. It is
hard for anyone to wrap their heads around everything that the
Coordinator tries to do. This refactor attempts to make interacting
with the Coordinator easier by separating it into logical parts. This
refactor is not perfect, but hopefully is a step in the right
direction.

There are two types of modules in this commit:
- Message Flow Modules: these modules contain the logic for how a
message is received, processed, and executed by the Coordinator.
- Utility Modules: These modules contain the logic and APIs for various
responsibilities and functionalities taken on by the Coordinator.
Unfortunately there is a gross interdependency between many of these
modules. I believe with some work we can tease out many of these
interdependencies.

An unfortunate side-effect of this is that the git history will become
very difficult to use for coord.rs.

This commit only does the following:
- Create new files.
- Rename files.
- Move methods/functions to a new file.
- Extract a block of logic into a function/method.
- Rename a function/method.
- Add doc comments.

Importantly, none of the actual logic has changed.