compute: truncate logging collections on restart
When computed restarts and continues writing its introspection data to
existing storage collections, it needs to ensure that they are initially
empty. Otherwise, we end up with duplicates that never get cleaned up
and produce confusing output when users query the introspection sources.
This commit implements the cleanup by reading a snapshot from the
storage collection and immediately retracting it.

Note that this does not protect from the scenario where an old
incarnation of the replica is still alive and also writing to the same
storage collections.