run.py: limit process killing on reset to current user's processes
This avoids crashing when there are storaged and computed processes
running in Docker; the user running run.py almost certainly is not
running as root and so can't kill the processes in Docker. The processes
in Docker are containerized, though, and therefore won't fight with the
storaged and computed processes spun up by run.py.