persist: track how many blob deletes are no-ops
Manually verified that this makes the bug fixed in #13953 immediately
obvious.

    # HELP mz_persist_external_blob_delete_noop_count count of blob delete calls that deleted a non-existent key
    # TYPE mz_persist_external_blob_delete_noop_count counter
    mz_persist_external_blob_delete_noop_count 53
    # HELP mz_persist_external_started_count count of external service calls started
    # TYPE mz_persist_external_started_count counter
    mz_persist_external_started_count{op="blob_delete"} 53

Because the s3 DeleteObject response doesn't indicate if it did any
work, we have to add a (slightly race-y but it's okay) HeadObject call
first. Since we have that, also return the number of deleted bytes so we
can track it in metrics.