persist: be more lenient about sinces when applying compaction results
It is perfectly safe, if the sinces don't match, as long as Spine
metadata's since is in advance of the one in the compaction result. This
mismatch can happen when the Spine has to be reloaded from state
(because of a compare_and_set mismatch).

This might also help an issue we're seeing in prod where compaction is
taking far longer than my mental model says it should.

Touches #13861