coord: optimize drop_temp_items
Don't execute catalog_transact with an empty ops. Putting this check
right in catalog_transact is a bit questionable because it has other
guarantees. Drop temp items is run on each session termination, so it
makes sense to optimize specifically that code path instead of changing
the semantics of catalog_transact.