persist: collect some shard level metrics
Ideally, we'd break these down per-shard, but that might be too
expensive for prometheus. Instead, start with the plumbing and aggregate
the interesting per-shard metrics across a process (min of since, max of
upper, sum of state size bytes). It's not ideal, but it's definitely
safe to ship.

The structure of the code makes it pretty trivial to do this swap
(actually tracking them per-shard) later: replace the ComputedFooGauge
in ShardsMetrics with a FooCounterVec and the AtomicFoo in ShardMetrics
with a FooCounter.

We hand out `Arc<ShardMetrics>` to read and write handles, but store it
here as `Weak`. This allows us to discover if it's no longer in use and
so we can remove it from the map.