coord: Add system table writes to group commit (#13960)
This commit updates system table writes so that they automatically
trigger a group commit and the system table writes are included in the
group commit. This simplifies the Coordinator slightly because all
non-empty writes are now processed in group commit, instead of two
separate places. Additionally, it allows any pending writes to user
tables to piggyback off of a write to a system table and be include
in a single timestamp and append call.