persist: assign gc work preferentially to writers
I originally thought it made sense to assign gc work proportionally to
whoever caused the seqno_since to move forward (either the last readers
to stop using a given seqno, or if there are no readers, any writer).
This has turned out to be wrong for at least two reasons:

- I was imagining that it would usually be some computed persist_source
  that would be selected. This might be the case for a big snapshot at
  computed startup, but in practice the steady-state case has turned out
  to be that this is some controller (enforcing our since compaction
  policy) and so _all_ gc runs in environmentd. The GC traffic for an
  env scales with the total persist traffic, so this is a pretty bad
  scaling bottleneck. Plus, environmentd is the most latency sensitive
  of any mz process. Fixing this is the immediate motivation for this
  PR.
- Even if the above hadn't been the case, we'd have continually selected
  whichever persist_source was most behind, potentially keeping it
  behind by adding additional (background) work.

Instead, Aljoscha suggested making GC, like compaction, the
responsibility of writers. Paul points out that we still want GC to
happen even if some storaged is down for a source while a potentially
large number of computed processes are calling downgrade_since, so this
probably means "preferentially writers". He also points out that we
happen to know in state whether there are any writers. It takes a
village to raise a GC policy.

So, now GC works as follows:

 - A new last_seq_req is maintained in State.
 - Whenever `apply_unbatched_cmd` computes a new_state for some cmd, it
   after checks if this should result in a GC req. If so, this is noted
   in new_state before we CaS (this allows us to sneak last_seq_req
   maintenance in without adding any new state versions).
 - This logic compares the current seqno_since of State to the last
   request against some threshold (to rate-limit GC a bit). It also only
   generates a request if the current cmd is a writer or if there are no
   writers registered.

In practice, this means in the common case user table GC still falls on
environmentd, but source GC happens in storaged and materialized view GC
happens in computed. If the uncommon case, someone still picks it up.