persist: check if state is from the future when decoding it
If persist gets some encoded ProtoState from the future (e.g. two
versions of code are running simultaneously against the same shard), it
might have a field that the current code doesn't know about. This would
be silently discarded at proto decode time. Unknown Fields [1] are a
tool we can use in the future to help deal with this, but in the
short-term, it's best to keep the persist read-modify-CaS loop simple
for as long as we can get away with it (i.e. until we have to offer the
ability to do rollbacks).

[1] https://developers.google.com/protocol-buffers/docs/proto3#unknowns

To detect the bad situation and disallow it, we tag every version of
state written to consensus with the version of code used to encode it.
Then at decode time, we're able to compare the current version against
any we receive and assert as necessary. Initially we reject any version
from the future (most conservative but easiest to reason about). If/when
we support deploy rollbacks and rolling upgrades, we can adjust this
assert as necessary to reflect the policy.

We could do the same for blob data, but it shouldn't be necessary. Any
blob data we read is going to be because we fetched it using a pointer
stored in some persist state. If we can handle the state, we can handle
the blobs it references, too.