pgtest: test invalid param binding
This fixes the test_bind_params flake.

Testing this with tokio_postgres as the driver causes the *driver*
to return the WrongType error message. Wireshark showed that the bind
request was never issued to the server. The test flake must be happening
because the driver's type cache either is or isn't yet populated, and
so it does end up issuing the bind request to the server which returns
the no overload error.