persist: remove API-level tracing events
It's never been clear if these were useful and with the tracing span
coverage, they're now redundant: we have tracing spans covering all the
same places now and tracing-subscriber has the ability to emit span
ENTER events as formatted logs. This also could be used to emit EXIT
events (which would correspond to the method finishing) and so is
strictly more general than this manual event start logging.

https://docs.rs/tracing-subscriber/0.3.14/tracing_subscriber/fmt/struct.SubscriberBuilder.html#method.with_span_events